+++ xparam/dup_tab.c
@@ -12,10 +12,10 @@
 
 #include "libft.h"
 
-int	*ft_linedup(int **tab, int size, int line)
+uint8_t	*ft_linedup(uint8_t **tab, uint8_t size, uint8_t line)
 {
-	int	i;
-	int	*tabline;
+	uint8_t	i;
+	uint8_t	*tabline;
 
 	i = 0;
 	tabline = malloc(sizeof(int) * size);
@@ -29,9 +29,9 @@ int	*ft_linedup(int **tab, int size, int line)
 	return (tabline);
 }
 
-int	**ft_freedup(int **tab, int j)
+uint8_t	**ft_freedup(uint8_t **tab, uint8_t j)
 {
-	int	i;
+	uint8_t	i;
 
 	i = 0;
 	while (i <= j)
@@ -42,14 +42,14 @@ int	**ft_freedup(int **tab, int j)
 	return (0);
 }
 
-int	**ft_tabdup(int **tab, int size)
+uint8_t	**ft_tabdup(uint8_t **tab, uint8_t size)
 {
-	int	j;
-	int	**duptab;
+	uint8_t	j;
+	uint8_t	**duptab;
 
 	j = 0;
 	size = size / 8 + 2;
-	duptab = malloc(sizeof(int *) * size);
+	duptab = malloc(sizeof(uint8_t *) * size);
 	if (duptab == 0)
 		return (0);
 	while (j < size)
